Migrate to postgres

+++ api/dbmodels/client.py
@@ -11,7 +11,7 @@ class Client(db.Model, Serializer):
 
     # Identification
     id = db.Column(db.Integer, primary_key=True)
-    user_id = db.Column(db.Integer, db.ForeignKey('user.id'), nullable=True)
+    user_id = db.Column(db.BigInteger, db.ForeignKey('user.id'), nullable=True)
     discord_user_id = db.Column(db.Integer, db.ForeignKey('discorduser.id'), nullable=True)
 
     # User Information
